Add all the things
Just playing around with gravity and the game loop. Will need to
restructure almost everything later.
